processor/requirements.txt
line-bot-sdk >= 2.4.2, < 3.0
boto3 >= 1.26.95, < 2.0
openai >= 0.27.2, < 1.0
jsonschema >= 4.17.3, < 5.0
tiktoken >= 0.3.2, < 1.0
